Replaced row/col based addressing by linear addressing
* Any nested row/col loops have been replaced by range based for loops
* compatibility methods have been added and the old row/col based access methods have been deprecated
* both versions of the onKeyswitchEvent handler are now used in parallel
* the old version of onKeyswitchEvent has been deprecated
* the KeyAddr class' toInt(), row() and col() methods are now used for indexed or row/col access to key addresses

+++ src/kaleidoscope/plugin/LEDEffect-Chase.h
@@ -58,7 +58,7 @@ class LEDChaseEffect : public Plugin,
 
     const LEDChaseEffect *parent_;
 
-    int8_t pos_ = 0;
+    int8_t pos_ = uint8_t(0);
     int8_t direction_ = 1;
     uint16_t last_update_ = 0;
   };
